dataset_harvester/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.3
pandas>=2.2.2
PyYAML>=6.0.1
tenacity>=8.2.3
playwright>=1.52.0
trafilatura>=1.9.0
lxml>=5.2.2
